refactor expression parsing into a helper
The code for parse_term and parse_expression was identical except for types, so now this is refactored into a helper function called parse_expression_level that uses a new templated type, AstExpressionLevel<TOperator, TInner>.

The new type describes an expression parsing precedence level and includes the enum of the operators that can be found at that level and which inner expression type is used. At the lowest level is parse_factor, which is hand-implemented.

The helper also relies on dedicated functions to parse TOperator.