Support for, while, do-while loops, break, and continue
This is the code to support for loops with and without a variable declaration, while loops, and do-while loops. It also supports the break and continue statements.

As part of this it adds support for empty statements so that the for loop header can have an empty initialization section.

To support break and continue, the codegen block state now includes the narrowest loop break and continue jump labels, if present.