Function definitions, function calls, and function parameters
This large change adds basic support for functions. The compiler now handles function declarations, definitions, multiple parameters, and function calls. It validates no duplicate function definitions and that all function calls have the correct number of parameters.

Implementing function calls is a significant part of the code because it involves implementing at least a part of the x64 calling convention. Arguments are passed in registers and on the stack. In particular, this compiler doesn't emit code that uses the base pointer (rbp). Instead it calculates the frame size and allocates all locations as offsets from the stack pointer (rsp).

To determine the correct frame size, the code runs the code generation part twice, but in the first pass it just records how much stack space was needed. In the second pass, it sets the stack space and correct offsets for locations.