Implement switch statements
This required adding a lot of semantic analysis stuff, in particular compile-time evaluation of constant expressions. That's needed for detecting duplicate cases in a switch.

Fixes a test issue where temp directories would be left over after running any test with an expected failure. Now temp directories are cleaned up automatically if the compilation was expected to fail.