Propagate better parse errors
Previously, basically every failure during the parsing stage ended up as the error "extra tokens after program end". It turns out I wasn't propagating the right errors from the parsing process.

I've also refactored the parsing code somewhat. Now several functions have a "try" mode, where they return Option<Type> instead of Result<Type> to indicate that the type of construct wasn't found but it wasn't an *error*.

Some of these helper functions now have a non-failing part and a failing part. For example, it's not a failure to not find the "for" keyword when parsing a new statement; it might not be a for loop next. But if the "for" keyword *is* found, then failure to parse anything after does constitute a failure: it's a malformed for loop.