Adds support for functions and function calls for Windows x64
This change rewrites a large amount of the compiler internals while also adding support for function declarations, definitions, and calls.

The biggest internal change that doesn't itself add a new feature is a pattern for traversing the different trees that the compiler maintains, in particular the AST and the ASM tree.

The new pattern introduces a function called `traverse_ast`, a `TraversalContext` object that goes with it, a callback function that can be used with it, and a `Nestable` trait that can be used to maintain a nesting structure during traversal.

This pattern simplifies some of the code that was there before like maintaining variable maps per scope and resolving all pseudoregisters.

This also introduces the ability to compile-only without linking (the `-c` command line arg) flag. And it introduces the ability to link against pre-compiled object files (the `--link` command line arg).

In order to link against externally-compiled C static libraries and to produce C static libraries that can be linked with externally-compiled executables, some of the ASM generation had to change to use some name mangling and other tricks, to work around the MASM assembler's goofy restrictions.

For debugging, you can now pass `-k` to keep intermediate files.

The tests got a few new abilities. Tests can now verify the stdout contents, in addition to the exit code.

`test_library` allows compiling two pieces of code: one the executable and the other the tests. One is compiled with this compiler and the other is compiled with MSVC. The two are linked together, the resultant binary executed, and the test checks for expected exit code and output. Then the compiler roles are switched. This helps test for a correctly implemented ABI.

It's not patterned yet, but one more test also links in an object from hand-written ASM and links it in to be callable during the test.