Move tests into individual files
This change removes all the tests from tests.rs and instead puts each one in an individual .c file, sorted by feature area.

There are different types of tests:
- `fail` - expect compilation failure
- `success` - compile, run, and check exit code
- `success_and_output` - same as success, but also check stdout
- `library` - compile a library and client code--one with MSVC and one with lrncc--link them together, and check expected exit code and optionally output. Then swap and build with the opposite compiler configuration

The .c files are converted each into a snippet of Rust code that runs the test from the file location, and all of that is included into tests.rs, which gives a hierarchical structure of modules mapped from directory paths.

Along the way, needed to fix a couple bugs that were missed because of test bugs in the old tests:
- don't permit duplicate parameter names in function declarations
- don't permit a trailing comma in the function parameter list